zn - refactoring two controllers and fix the test
